+++ packages/observer-engine/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @perceo/observer-engine
 
+## 3.0.1
+
+### Patch Changes
+
+- Updating logginf or the temporal and cli output off init
+- Updated dependencies
+  - @perceo/supabase@0.3.1
+  - @perceo/perceo@0.4.1
+
+## 3.0.0
+
+### Minor Changes
+
+- Completely revamped the observer engine's init/bootstrap workflow to better suit how I want flows to be generated/how I want them to look
+
+### Patch Changes
+
+- Updated dependencies
+  - @perceo/supabase@0.3.0
+  - @perceo/perceo@0.4.0
+
+## 2.0.5
+
+### Patch Changes
+
+- Try to fix the .env not working again.
+- Updated dependencies
+  - @perceo/supabase@0.2.4
+  - @perceo/perceo@0.3.5
+
+## 2.0.4
+
+### Patch Changes
+
+- Fix anon key and supabase url not present for login before pulling public_env
+- Updated dependencies
+  - @perceo/supabase@0.2.3
+  - @perceo/perceo@0.3.4
+
+## 2.0.3
+
+### Patch Changes
+
+- Adding loading for .env variables from public env
+- Updated dependencies
+  - @perceo/supabase@0.2.2
+  - @perceo/perceo@0.3.3
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @perceo/supabase@0.2.1
+  - @perceo/perceo@0.3.2
+
 ## 2.0.1
 
 ### Patch Changes