+++ packages/supabase/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @perceo/supabase
 
+## 0.3.1
+
+### Patch Changes
+
+- Updating logginf or the temporal and cli output off init
+
+## 0.3.0
+
+### Minor Changes
+
+- Completely revamped the observer engine's init/bootstrap workflow to better suit how I want flows to be generated/how I want them to look
+
+## 0.2.4
+
+### Patch Changes
+
+- Try to fix the .env not working again.
+
+## 0.2.3
+
+### Patch Changes
+
+- Fix anon key and supabase url not present for login before pulling public_env
+
+## 0.2.2
+
+### Patch Changes
+
+- Adding loading for .env variables from public env
+
+## 0.2.1
+
+### Patch Changes
+
+- Fix a minor rpc error
+
 ## 0.2.0
 
 ### Minor Changes